refactor: consolidate market data module
